* master: (83 commits)
  CC-1177: Fixed CompanyPage controllers
  CC-1197: Fixed language "switcher.twig"
  CC-796: Updated "composer.json" dependencies for modules with new widgets
  CC-796: Updated template "product-configurator"
  CC-796: Reverting deletion of "DismissController" and template "dismiss-confirm"
  CC-796: Updated "composer.json" files for modules with new global widgets usage in templates
  CC-796: Fixed "TODO"s
  CC-796: Fixed "composer.json" for modules with new global widgets
  CC-796: Fixed "README.md" for module "ProductBarcodeWidget"
  CC-796: Fixed "README.md" for module "ProductBarcodeWidget"
  PS-5016: fix PR comment, replcase comparing with false to 'not' constructions
  CC-796: Fixed "README.md" for module "ProductBarcodeWidget"
  CC-796: Fixed "README.md" for module "ProductBarcodeWidget"
  CC-796: Fixed errors found with code sniffer and php stan
  CC-796: Fixed errors found with code sniffer and php stan
  CC-796: Added descriptions in "README.md" for several modules
  CC-796: Changed templates accordingly to CR comments
  CC-796: Changes for "ProductSetWidget" templates accordingly to CR comments
  CC-796: Fixed conflicts of merge epic master into epic dev
  CC-796: Updated templates and widget plugins
  ...

+++ Bundles/CompanyPage/src/SprykerShop/Yves/CompanyPage/Controller/CompanyRoleController.php
@@ -12,11 +12,12 @@ use Generated\Shared\Transfer\CompanyRoleCriteriaFilterTransfer;
 use Generated\Shared\Transfer\CompanyRoleResponseTransfer;
 use Generated\Shared\Transfer\CompanyRoleTransfer;
 use Generated\Shared\Transfer\CompanyUserCollectionTransfer;
-use Generated\Shared\Transfer\CompanyUserTransfer;
 use Generated\Shared\Transfer\PermissionCollectionTransfer;
+use SprykerShop\Yves\CompanyPage\Form\CompanyRoleForm;
 use SprykerShop\Yves\CompanyPage\Plugin\Provider\CompanyPageControllerProvider;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class CompanyRoleController extends AbstractCompanyController
 {
@@ -73,6 +74,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return array
      */
     protected function executeDetailsAction(Request $request): array
@@ -83,6 +86,10 @@ class CompanyRoleController extends AbstractCompanyController
         $companyRoleTransfer = $this->getFactory()
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $companyRolePermissions = $this->getFactory()
             ->getCompanyRoleClient()
@@ -101,6 +108,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return \Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function deleteAction(Request $request): RedirectResponse
@@ -108,6 +117,14 @@ class CompanyRoleController extends AbstractCompanyController
         $idCompanyRole = $request->query->getInt('id');
         $companyRoleTransfer = new CompanyRoleTransfer();
         $companyRoleTransfer->setIdCompanyRole($idCompanyRole);
+
+        $companyRoleTransfer = $this->getFactory()
+            ->getCompanyRoleClient()
+            ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $companyRoleResponseTransfer = $this->getFactory()->getCompanyRoleClient()->deleteCompanyRole($companyRoleTransfer);
 
@@ -123,6 +140,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return \Spryker\Yves\Kernel\View\View|\Symfony\Component\HttpFoundation\RedirectResponse
      */
     public function confirmDeleteAction(Request $request)
@@ -136,6 +155,10 @@ class CompanyRoleController extends AbstractCompanyController
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
 
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
+
         if ($companyRoleTransfer->getIsDefault()) {
             $this->addErrorMessage(static::ERROR_MESSAGE_DEFAULT_COMPANY_ROLE_DELETE);
 
@@ -147,6 +170,8 @@ class CompanyRoleController extends AbstractCompanyController
 
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
+     *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
      *
      * @return \Spryker\Yves\Kernel\View\View
      */
@@ -160,6 +185,10 @@ class CompanyRoleController extends AbstractCompanyController
         $companyRoleTransfer = $this->getFactory()
             ->getCompanyRoleClient()
             ->getCompanyRoleById($companyRoleTransfer);
+
+        if (!$this->isCurrentCustomerRelatedToCompany($companyRoleTransfer->getFkCompany())) {
+            throw new NotFoundHttpException();
+        }
 
         $viewData = [
             'idCompanyRole' => $idCompanyRole,
@@ -203,7 +232,7 @@ class CompanyRoleController extends AbstractCompanyController
             ->handleRequest($request);
 
         if ($companyRoleForm->isSubmitted() === false) {
-            $idCompany = $this->getCompanyUser()->getFkCompany();
+            $idCompany = $this->findCurrentCompanyUserTransfer()->getFkCompany();
             $companyRoleForm->setData($dataProvider->getData($idCompany));
         }
 
@@ -242,6 +271,8 @@ class CompanyRoleController extends AbstractCompanyController
     /**
      * @param \Symfony\Component\HttpFoundation\Request $request
      *
+     * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
+     *
      * @return array|\Symfony\Component\HttpFoundation\RedirectResponse
      */
     protected function executeUpdateAction(Request $request)
@@ -259,8 +290,13 @@ class CompanyRoleController extends AbstractCompanyController
         $idCompanyRole = $request->query->getInt(static::PARAMETER_ID_COMPANY_ROLE);
 
         if ($companyRoleForm->isSubmitted() === false) {
-            $idCompany = $this->getCompanyUser()->getFkCompany();
-            $companyRoleForm->setData($dataProvider->getData($idCompany, $idCompanyRole));
+            $idCompany = $this->findCurrentCompanyUserTransfer()->getFkCompany();
+            $data = $dataProvider->getData($idCompany, $idCompanyRole);
+
+            if (!$this->isCurrentCustomerRelatedToCompany($data[CompanyRoleForm::FIELD_FK_COMPANY])) {
+                throw new NotFoundHttpException();
+            }
+            $companyRoleForm->setData($data);
         }
 
         if ($companyRoleForm->isSubmitted() && $companyRoleForm->isValid()) {
@@ -286,7 +322,7 @@ class CompanyRoleController extends AbstractCompanyController
     {
         $criteriaFilterTransfer = new CompanyRoleCriteriaFilterTransfer();
 
-        $criteriaFilterTransfer->setIdCompany($this->getCompanyUser()->getFkCompany());
+        $criteriaFilterTransfer->setIdCompany($this->findCurrentCompanyUserTransfer()->getFkCompany());
         $criteriaFilterTransfer->setPagination($this->createPaginationTransfer($request));
         $criteriaFilterTransfer->setFilter($this->createFilterTransfer(static::COMPANY_ROLE_SORT_FIELD));
 
@@ -359,21 +395,5 @@ class CompanyRoleController extends AbstractCompanyController
         return $this->getFactory()
             ->getCompanyRoleClient()
             ->findNonInfrastructuralCompanyRolePermissionsByIdCompanyRole($companyRoleTransfer);
-    }
-
-    /**
-     * @return \Generated\Shared\Transfer\CompanyUserTransfer|null
-     */
-    protected function findCurrentCompanyUserTransfer(): ?CompanyUserTransfer
-    {
-        $currentCustomerTransfer = $this->getFactory()
-            ->getCustomerClient()
-            ->getCustomer();
-
-        if (!$currentCustomerTransfer) {
-            return null;
-        }
-
-        return $currentCustomerTransfer->getCompanyUserTransfer();
     }
 }
